requirements.txt
astroid==2.4.2
colorama==0.4.4
et-xmlfile==1.0.1
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.22
openpyxl==3.0.5
protobuf==3.13.0
pylint==2.6.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
